Improving the Matrix/maths part of the library.
Tested the conversion from euler, rotation matrix and quaternion.

Simplify the code to use heavily the eigen library.

Conversion are tested with data obtained from
a verified website, use by teachers at ETS and verified
by guys with PHD in maths, so it is ok.

Added a conversion module. For now only do radian-degree

Had to modify the CMakeList to link eigen lib.