Modification to the Euler angle handling
In test, we realized we had issue on the yaw, it was changing sign and value.
This is due to the order or rotation. For our quaternion, the rotation with
respect to the world is the yaw, not the pitch. Inversed the X and Z in the rotation

Now retursn yaw pitch roll.